requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ollama==0.5.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
